calendar_project/py35_requirements.txt
arrow==0.12.1
binaryornot==0.4.4
business-rules==1.0.1
certifi==2018.1.18
chardet==3.0.4
click==6.7
cookiecutter==1.6.0
Django==2.0.4
django-bootstrap4==0.0.6
django-filter==1.1.0
django-macros==0.4.0
django-mssql==1.8
django-nose==1.4.5
django-pyodbc==1.1.3
django-pyodbc-azure==2.0.4.0
django-templates-macros==0.2
django-webpack-loader==0.6.0
djangorestframework==3.7.7
djangorestframework-datatables==0.2.1
Flask==0.12.2
Flask-Restless==0.17.0
Flask-SQLAlchemy==2.3.2
future==0.16.0
gunicorn==19.7.1
idna==2.6
ipdb==0.8
ipython==1.2.1
itsdangerous==0.24
Jinja2==2.10
jinja2-time==0.2.0
Markdown==2.6.11
MarkupSafe==1.0
mimerender==0.6.0
mock==1.0.1
nose==1.3.1
nose-run-line-number==0.0.1
poyo==0.4.1
pyodbc==4.0.23
python-dateutil==2.7.2
python-mimeparse==1.6.0
pytz==2018.3
requests==2.18.4
six==1.11.0
SQLAlchemy==1.2.6
urllib3==1.22
virtualenv==15.2.0
Werkzeug==0.14.1
whichcraft==0.4.1
